requirements.txt
absl-py==2.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.8
easydict==1.12
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.19
urllib3==2.2.1
Werkzeug==3.0.1
